+++ src/main/java/sopt/org/hmh/domain/auth/service/AuthService.java
@@ -1,30 +1,18 @@
 package sopt.org.hmh.domain.auth.service;
 
-import java.time.LocalDateTime;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
+import sopt.org.hmh.domain.auth.dto.response.ReissueResponse;
 import sopt.org.hmh.domain.challenge.service.ChallengeService;
-import sopt.org.hmh.domain.users.domain.OnboardingInfo;
-import sopt.org.hmh.domain.users.domain.OnboardingProblem;
 import sopt.org.hmh.domain.users.domain.User;
-import sopt.org.hmh.domain.users.domain.UserConstants;
-import sopt.org.hmh.domain.auth.exception.AuthError;
-import sopt.org.hmh.domain.auth.exception.AuthException;
 import sopt.org.hmh.domain.auth.dto.request.SocialPlatformRequest;
 import sopt.org.hmh.domain.auth.dto.request.SocialSignUpRequest;
 import sopt.org.hmh.domain.auth.dto.response.LoginResponse;
-import sopt.org.hmh.domain.auth.dto.response.ReissueResponse;
-import sopt.org.hmh.domain.auth.repository.OnboardingInfoRepository;
-import sopt.org.hmh.domain.auth.repository.ProblemRepository;
-import sopt.org.hmh.domain.users.repository.UserRepository;
-import sopt.org.hmh.global.auth.jwt.JwtProvider;
-import sopt.org.hmh.global.auth.jwt.JwtValidator;
+import sopt.org.hmh.domain.users.service.UserService;
+import sopt.org.hmh.global.auth.jwt.TokenService;
 import sopt.org.hmh.global.auth.jwt.exception.JwtError;
 import sopt.org.hmh.global.auth.jwt.exception.JwtException;
-import sopt.org.hmh.global.auth.redis.TokenService;
 import sopt.org.hmh.global.auth.social.SocialPlatform;
 import sopt.org.hmh.global.auth.social.SocialAccessTokenResponse;
 import sopt.org.hmh.global.auth.social.apple.fegin.AppleOAuthProvider;
@@ -35,15 +23,12 @@ import sopt.org.hmh.global.auth.social.kakao.fegin.KakaoLoginService;
 @Transactional(readOnly = true)
 public class AuthService {
 
-    private final JwtProvider jwtProvider;
-    private final JwtValidator jwtValidator;
-    private final UserRepository userRepository;
-    private final OnboardingInfoRepository onboardingInfoRepository;
-    private final ProblemRepository problemRepository;
     private final KakaoLoginService kakaoLoginService;
+    private final AppleOAuthProvider appleOAuthProvider;
+
     private final ChallengeService challengeService;
     private final TokenService tokenService;
-    private final AppleOAuthProvider appleOAuthProvider;
+    private final UserService userService;
 
     @Transactional
     public LoginResponse login(String socialAccessToken, SocialPlatformRequest request) {
@@ -51,7 +36,7 @@ public class AuthService {
         SocialPlatform socialPlatform = request.socialPlatform();
         String socialId = getSocialIdBySocialAccessToken(socialPlatform, socialAccessToken);
 
-        User loginUser = getUserBySocialPlatformAndSocialId(socialPlatform, socialId);
+        User loginUser = userService.getUserBySocialPlatformAndSocialId(socialPlatform, socialId);
 
         return performLogin(socialAccessToken, socialPlatform, loginUser);
     }
@@ -62,9 +47,9 @@ public class AuthService {
         SocialPlatform socialPlatform = request.socialPlatform();
         String socialId = getSocialIdBySocialAccessToken(socialPlatform, socialAccessToken);
 
-        validateDuplicateUser(socialId, socialPlatform);
+        userService.validateDuplicateUser(socialId, socialPlatform);
 
-        User user = addUser(socialPlatform, socialId, request.name());
+        User user = userService.addUser(socialPlatform, socialId, request.name());
 
         challengeService.updateChallengeForPeriodWithInfo(
                 challengeService.addChallenge(user.getId(),
@@ -72,42 +57,9 @@ public class AuthService {
                         request.challengeSignUpRequest().goalTime()),
                 request.challengeSignUpRequest().apps(),
                 os);
-        registerOnboardingInfo(request);
+        userService.registerOnboardingInfo(request);
 
         return performLogin(socialAccessToken, socialPlatform, user);
-    }
-
-    @Transactional
-    public ReissueResponse reissueToken(String refreshToken) {
-        refreshToken = parseTokenString(refreshToken);
-        Long userId = jwtProvider.getSubject(refreshToken);
-        validateRefreshToken(refreshToken, userId);
-        tokenService.deleteRefreshToken(userId);
-        return ReissueResponse.of(jwtProvider.issueToken(userId));
-    }
-
-    private void validateUserId(Long userId) {
-        if (!userRepository.existsById(userId)) {
-            throw new AuthException(AuthError.NOT_FOUND_USER);
-        }
-    }
-
-    private void validateRefreshToken(String refreshToken, Long userId) {
-        try {
-            jwtValidator.validateRefreshToken(refreshToken);
-            validateUserId(userId);
-        } catch (JwtException jwtException) {
-            tokenService.deleteRefreshToken(userId);
-            throw jwtException;
-        }
-    }
-
-    private User getUserBySocialPlatformAndSocialId(SocialPlatform socialPlatform, String socialId) {
-        User user = userRepository.findBySocialPlatformAndSocialIdOrThrowException(socialPlatform, socialId);
-        if (user.isDeleted()) {
-            user.recover();
-        }
-        return user;
     }
 
     private String getSocialIdBySocialAccessToken(SocialPlatform socialPlatform, String socialAccessToken) {
@@ -118,62 +70,17 @@ public class AuthService {
         };
     }
 
-    /**
-     * 소셜 액세스 토큰에서 "Bearer " 부분을 삭제시키고 유효한 소셜 액세스 토큰만을 받기 위한 함수
-     */
-    private String parseTokenString(String tokenString) {
-        String[] parsedTokens = tokenString.split(" ");
-        if (parsedTokens.length != 2) {
-            throw new JwtException(JwtError.INVALID_TOKEN_HEADER);
-        }
-        return parsedTokens[1];
+    public ReissueResponse reissueToken(String refreshToken) {
+        return tokenService.reissueToken(refreshToken);
     }
 
-    private void validateDuplicateUser(String socialId, SocialPlatform socialPlatform) {
-        if (userRepository.existsBySocialPlatformAndSocialId(socialPlatform, socialId)) {
-            throw new AuthException(AuthError.DUPLICATE_USER);
-        }
-    }
 
     private LoginResponse performLogin(String socialAccessToken, SocialPlatform socialPlatform, User loginUser) {
         if (socialPlatform == SocialPlatform.KAKAO) {
             kakaoLoginService.updateUserInfoByKakao(loginUser, socialAccessToken);
         }
-        return LoginResponse.of(loginUser, jwtProvider.issueToken(loginUser.getId()));
+        return LoginResponse.of(loginUser, tokenService.issueToken(loginUser.getId()));
     }
-
-    private User addUser(SocialPlatform socialPlatform, String socialId, String name) {
-        return userRepository.save(
-                User.builder()
-                        .socialPlatform(socialPlatform)
-                        .socialId(socialId)
-                        .name(validateName(name))
-                        .build()
-        );
-    }
-
-    private String validateName(String name) {
-        if (!StringUtils.hasText(name)) {
-            return UserConstants.DEFAULT_USER_NAME;
-        }
-        return name;
-    }
-
-    private void registerOnboardingInfo(SocialSignUpRequest request) {
-        OnboardingInfo onboardingInfo = OnboardingInfo.builder()
-                .averageUseTime(request.onboardingRequest().averageUseTime())
-                .build();
-        Long onboardingInfoId = onboardingInfoRepository.save(onboardingInfo).getId();
-
-        List<OnboardingProblem> problemList = request.onboardingRequest().problemList().stream()
-                .map(problem -> OnboardingProblem.builder()
-                        .onboardingInfoId(onboardingInfoId)
-                        .problem(problem).build())
-                .toList();
-        problemRepository.saveAll(problemList);
-    }
-
-
 
     public SocialAccessTokenResponse getSocialAccessTokenByAuthorizationCode(String code) {
         return kakaoLoginService.getKakaoAccessToken(code);
